submissions/submission-89/pyproject.toml
[project]
name = "gmm-v4"
version = "0.1.0"
description = "Gaussian Mixture Models v4"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.8"
dependencies = [
    "torch",
    "numpy",
    "matplotlib",
    "pandas",
    "tqdm",
    "pillow",
    "ipython",
    "scipy",
    "scikit-learn",
    "pydantic",
    "jax",
    "ott-jax",
    "tensorflow",
    "tensorboard",
    "POT",
    "fast-pytorch-kmeans",
    "sympy",
    "networkx",
    "requests",
    "h5py",
    "pyarrow",
    "ipykernel>=6.29.5",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "model",
    "utils", 
    "training",
    "losses",
    "metrics",
    "mi",
    "config"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.991",
] 
